requirements.txt
fastapi
uvicorn
python-multipart
gunicorn
requests
flask
flask-cors
yfinance
openai
GoogleNews
duckduckgo-search
praw
python-dotenv
vaderSentiment
pypdf
youtube-transcript-api
